+++ app/controllers/dmptool/org_admin/templates_controller.rb
@@ -27,7 +27,83 @@ module Dmptool
 
         render '/org_admin/templates/email' # .js.erb'
       end
+      # rubocop:enable Metrics/AbcSize
+
+      def preferences
+        template = Template.find(params[:id])
+        authorize Template
+        render 'preferences', locals: {
+          partial_path: 'edit',
+          template: template,
+          output_types: ResearchOutput.output_types,
+          preferred_licenses: License.preferred.map { |license| [license.identifier, license.id] },
+          licenses: License.selectable.map { |license| [license.identifier, license.id] }
+        }
+      end
+
+      # GET /org_admin/templates/[:id] # ,
+      # rubocop:disable Metrics/AbcSize
+      def save_preferences
+        template = Template.find(params[:id])
+        authorize Template
+
+        args = preference_params
+        args[:customize_output_types] = params[:customize_output_types_sel] != '0'
+        args[:customize_licenses] = params[:customize_licenses_sel] != '0'
+        Template.transaction do
+          template.update(template_output_types: [], licenses: [], repositories: [], metadata_standards: [])
+          template.update(args)
+          template.update(repositories: []) if preference_params[:customize_repositories] == '0'
+          template.update(metadata_standards: []) if preference_params[:customize_metadata_standards] == '0'
+        end
+        preferences
+      end
+      # rubocop:enable Metrics/AbcSize
+
+      # rubocop:disable Metrics/AbcSize
+      def repository_search
+        @template = Template.find(params[:id])
+        authorize Template
+
+        @search_results = Repository.by_type(repo_search_params[:type_filter])
+        @search_results = @search_results.by_subject(repo_search_params[:subject_filter])
+        @search_results = @search_results.search(repo_search_params[:search_term])
+
+        @search_results = @search_results.order(:name).page(params[:page])
+      end
+      # rubocop:enable Metrics/AbcSize
+
+      def metadata_standard_search
+        @template = Template.find(params[:id])
+        authorize Template
+
+        @search_results = MetadataStandard.search(metadata_standard_search_params[:search_term])
+                                          .order(:title)
+                                          .page(params[:page])
+      end
+
+      private
+
+      def repo_search_params
+        params.require(:template).permit(%i[search_term subject_filter type_filter])
+      end
+
+      def metadata_standard_search_params
+        params.require(:template).permit(%i[search_term])
+      end
+
+      def preference_params
+        params.require(:template).permit(
+          :enable_research_outputs,
+          :user_guidance_output_types, :user_guidance_repositories,
+          :user_guidance_output_types_title, :user_guidance_output_types_description,
+          :user_guidance_metadata_standards, :user_guidance_licenses,
+          :customize_output_types, :customize_repositories,
+          :customize_metadata_standards, :customize_licenses,
+          template_output_types_attributes: %i[id research_output_type],
+          licenses_attributes: %i[id], repositories_attributes: %i[id], metadata_standards_attributes: %i[id]
+        )
+      end
     end
-    # rubocop:enable Metrics/AbcSize
   end
 end