+++ app/javascript/src/dmptool/orgAdmin/templates/preferences.js
@@ -1,6 +1,29 @@
 import { Tinymce } from '../../../utils/tinymce';
 
 $(() => {
+  if ($('#template_user_guidance_repositories').length > 0) {
+    Tinymce.init({ selector: '#template_user_guidance_repositories' });
+  }
+  if ($('#template_user_guidance_metadata_standards').length > 0) {
+    Tinymce.init({ selector: '#template_user_guidance_metadata_standards' });
+  }
+});
+
+$(() => {
+  function setPrefsControls() {
+    if ($('#template_enable_research_outputs:checked').is('*')) {
+      $('h2.prefs_option, div.prefs_option').show();
+    } else {
+      $('h2.prefs_option, div.prefs_option').hide();
+    }
+  }
+  $('#template_enable_research_outputs').on('click', () => {
+    setPrefsControls();
+  });
+  setPrefsControls();
+});
+
+$(() => {
   $('a.output_type_remove').on('click', (e) => {
     e.stopPropagation();
     $(e.currentTarget).parents('li.output_type').remove();
@@ -34,13 +57,13 @@ $(() => {
   function addOutputType(v) {
     const vnorm = v.replace(/^\s+|\s+$/g, '').toLowerCase();
     const vnormDisp = vnorm.charAt(0).toUpperCase() + vnorm.slice(1);
-    if (checkOutputType('#my-output-types', vnorm)) {
+    if (checkOutputType('#my-output-types', vnormDisp)) {
       return;
     }
-    const vclass = checkOutputType('#default-output-types', vnorm) ? 'standard' : 'custom';
+    const vclass = checkOutputType('#default-output-types', vnormDisp) ? 'standard' : 'custom';
     const li = $('<li/>').addClass('selectable_item').addClass('output_type').addClass(vclass)
       .appendTo('#my-output-types ul');
-    const a = $('<a href="#" aria-label="Remove this output type"/>').addClass('output_type_remove').appendTo(li);
+    const a = $('<a/>').attr('aria-label', "Remove this output type").addClass('output_type_remove').appendTo(li);
     a.on('click', (e) => {
       e.stopPropagation();
       $(e.currentTarget).parents('li.output_type').remove();
@@ -81,3 +104,150 @@ $(() => {
     return false;
   });
 });
+
+$(() => {
+  $('a.license_remove').on('click', (e) => {
+    e.stopPropagation();
+    $(e.currentTarget).parents('li.license').remove();
+  });
+
+  function showLicenseSelections() {
+    if ($('#customize_licenses_sel').val() === '0') {
+      $('#default-licenses').show();
+      $('#my-licenses').hide();
+      $('#my-licenses input.license').attr('disabled', true);
+    } else {
+      $('#default-licenses').hide();
+      $('#my-licenses').show();
+      $('#my-licenses input.license').attr('disabled', false);
+    }
+  }
+  showLicenseSelections();
+
+  function checkLicense(sel, v) {
+    let res = false;
+    const ns = $(sel).find('ul li.license');
+    ns.each((n) => {
+      const node = $(ns.get(n));
+      const cv = node.find('input.license').val();
+      if (v === cv) {
+        res = true;
+      }
+    });
+    return res;
+  }
+
+  function addLicense(id, v) {
+    if (checkLicense('#my-licenses', id)) {
+      return;
+    }
+    const vclass = checkLicense('#default-licenses', id) ? 'standard' : 'custom';
+    const li = $('<li/>').addClass('selectable_item').addClass('license').addClass(vclass)
+      .appendTo('#my-licenses ul');
+    const a = $('<a/>').attr('aria-label', "Remove this license").addClass('license_remove').appendTo(li);
+    a.on('click', (e) => {
+      e.stopPropagation();
+      $(e.currentTarget).parents('li.license').remove();
+    });
+    const span = $('<span/>').addClass('selectable_item_label').addClass(vclass).appendTo(a);
+    span.text(v);
+    const index = $('#my-licenses ul li').length;
+    $('<i class="fas fa-times-circle fa-reverse remove-license" aria-hidden="true"/>').appendTo(a);
+    const name = `template[licenses_attributes[${index}][id]]`;
+    $('<input class="license" type="hidden" autocomplete="off"/>').attr('name', name).val(id).appendTo(li);
+  }
+
+  $('input.license_init').each((n) => {
+    const node = $($('input.license_init').get(n));
+    addLicense(node.val(), node.attr('data'));
+  }).remove();
+
+  $('#customize_licenses_sel').on('change', (e) => {
+    e.stopPropagation();
+    if ($('#customize_licenses_sel').val() === '1') {
+      if ($('#my-licenses ul li').length === 0) {
+        $('#default-licenses ul li.license').each((n) => {
+          const node = $($('#default-licenses ul li.license').get(n));
+          addLicense(node.find('input.license').val(), node.find('input.license').attr('data'));
+        });
+      }
+    }
+    showLicenseSelections();
+  });
+
+  $('#add_license').on('click', () => {
+    const v = $('#new_license').val();
+    if (v !== '') {
+      addLicense(v, $('#new_license option:selected').text());
+    }
+    return false;
+  });
+});
+
+$(() => {
+  function setModalButtonRepo() {
+    if ($('#template_customize_repositories:checked').is('*')){
+      $('#prefs-repositories').show();
+    } else {
+      $('#prefs-repositories').hide();
+    }
+  }
+  $('#template_customize_repositories').on('change', (e) => {
+    setModalButtonRepo();
+  });
+  setModalButtonRepo();
+});
+
+$(() => {
+  function setModalButtonMetadata() {
+    if ($('#template_customize_metadata_standards:checked').is('*')){
+      $('#prefs-metadata_standards').show();
+    } else {
+      $('#prefs-metadata_standards').hide();
+    }
+  }
+  $('#template_customize_metadata_standards').on('change', (e) => {
+    setModalButtonMetadata();
+  });
+  setModalButtonMetadata();
+});
+
+$(() => {
+  function check(sel_enable, sel_count, msg) {
+    if ($(sel_enable).is("*")){
+      if ($(sel_count).length == 0) {
+        alert(msg);
+        return false;
+      }
+    }
+    return true;
+  } 
+
+  $('form.edit_template').on('submit', (e) => {
+    let b = true;
+    b = b & check(
+      "#customize_output_types_sel option[value!='0']:selected",
+      "input.output_type[type='hidden']:enabled",
+      "At least one preferred OUTPUT TYPE must be selected if you are enabling a preferred list."
+    );
+    b = b & check(
+      '#template_customize_repositories:checked',
+      '#modal-search-repositories-selections div.modal-search-result-label',
+      "At least one preferred REPOSITORY must be selected if you are enabling a preferred list."
+    );
+    b = b & check(
+      '#template_customize_metadata_standards:checked',
+      '#modal-search-metadata_standards-selections div.modal-search-result-label',
+      "At least one preferred METADATA STANDARD must be selected if you are enabling a preferred list."
+    );
+    b = b & check(
+      "#customize_licenses_sel option[value!='0']:selected",
+      "#my-licenses input.license[type='hidden']:enabled",
+      "At least one preferred LICENSE must be selected if you are enabling a preferred list."
+    );
+    if (!b) {
+      e.stopPropagation();
+      e.preventDefault();
+    }
+  })
+})