+++ db/migrate/20230329235100_end_user_guidance2.rb
@@ -1,0 +1,6 @@
+class EndUserGuidance2 < ActiveRecord::Migration[6.1]
+  def change
+    add_column :templates, :user_guidance_output_types_title, :text
+    add_column :templates, :user_guidance_output_types_description, :text
+  end
+end